Compatible with LuaJIT (using FFI structs), Lua 5.1 and Lua 5.4

Performance change compared to [brainfuck interpreter](https://github.com/kostya/benchmarks/blob/master/brainfuck/bf.lua)



LuaJIT :

	mandelbrot -> ran in 53.81s -> ran in 2.183s

	bench -> ran in 8.25s -> ran in 0.183s

**2364.96%** and **4408.2%** perf boost



Lua 51 : 

	bench -> ran in 98.926s -> ran in 14.042s

**604.5% perf boost**


Lua 54 :

	bench -> ran in 51.345s -> ran in 4.858s

**956.92%** perf boost


LuaJIT goes BRRRRRRR