@kostya 2020-11-01
You completely change algorithm, this is not accepted. We can do similar in other languages and have boost too.