@kostya 2020-11-01
we can use different implementation details and tunes, but algorithm should be the same, to be comparable between languages.