@ExtReMLapin 2020-11-03
Could this be accepted @kostya ?


```lua
local INC = 0
local MOVE = 1
local PRINT = 3
local LOOP = 4
local isJIT = type(jit) == "table"

local function Brainfuck(text)
	local pos = 1
	local data

	if isJIT then
		data = require("ffi").new("int[30000]") -- cf: wikipedia
	else
		data = {}
		local i = 0

		while (i < 30000) do
			data[i] = 0
			i = i + 1
		end
	end

	local insData = {
		["+"] = {INC, 1},
		["-"] = {INC, -1},
		[">"] = {MOVE, 1},
		["<"] = {MOVE, -1}
	}

	local function parse(source, i)
		local res = {}
		local len = source:len()

		while i <= len do
			local c = source:sub(i, i)
			local insPtr = insData[c]

			if insPtr then
				local count = 1

				while i <= len and source:sub(i + 1, i + 1) == c do
					count = count + 1
					i = i + 1
				end

				table.insert(res, {insPtr[1], insPtr[2] * count})
			elseif c == "." then
				table.insert(res, {PRINT})
			elseif c == "[" then
				local loop_code
				loop_code, i = parse(source, i + 1)
				table.insert(res, {LOOP, loop_code})
			elseif c == "]" then
				break
			end

			i = i + 1
		end

		return res, i
	end

	local function _run(program)
		for i = 1, #program do
			local op = program[i]
			local operator = op[1]

			if operator == INC then
				data[pos] = data[pos] + op[2]
			elseif operator == MOVE then
				pos = pos + op[2]
			elseif operator == LOOP then
				while data[pos] > 0 do
					_run(op[2], nil)
				end
			elseif operator == PRINT then
				io.write(string.char(data[pos]))
				io.flush()
			end
		end
	end

	local function run()
		_run(parse(text, 1))
	end

	return {
		run = run
	}
end
```