@kostya 2020-11-03
You remove Tape abstraction, are you think this is good for real tasks usage? Should not language inline it by itself? Also prefer to use autogrowing array, as all languages use.