@ExtReMLapin 2020-11-03
I removed Tape abstraction for personal readability.

The wikipedia article states it needs an array of 30k, shouldn't all language use it ?