@kostya 2020-11-03
We decide just use everywhere idiomatic array storage.