@nuald 2020-11-04
@ExtReMLapin The purpose of the bf benchmark is testing the performance of the idiomatic constructions, it could be any other general-purpose algorithm. However, I totally understand that you may want to have better, optimized version for the reference (in this particular example, FFI makes sense as Lua is intended to be embedded and communicate a lot with the native code). 

Therefore, my recommendation would be not changing the reference implementation (that should have the similar algorithm and the similar list/array operations), but add a new one and mark it as "Lua (FFI)" (I'm not sure FFI is a reason for the performance increase though, I'd leave the proper naming to you). You could see that we have the same with Haskell (reference implementation and MArray-based). If you have troubles adding it into the toolchain, just please submit a new file, and I'll add it to all required places (just let me know do you need it for Lua and/or LuaJIT).

Also, if you see some obvious mistakes in the reference implementation, you're welcome to address those unless algorithm changes are required.